chore: rename Maven coordinates to com.bolaneradarmini for project consistency
