+++ src/main/java/org/bolaneradarmini/config/SecurityConfig.java
@@ -1,4 +1,4 @@
-package org.example.config;
+package org.bolaneradarmini.config;
 
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
