+++ src/main/java/org/bolaneradarmini/controller/BankController.java
@@ -1,7 +1,7 @@
-package org.example.controller;
+package org.bolaneradarmini.controller;
 
-import org.example.entity.Bank;
-import org.example.repository.BankRepository;
+import org.bolaneradarmini.entity.Bank;
+import org.bolaneradarmini.repository.BankRepository;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
