+++ src/main/java/org/bolaneradarmini/entity/AverageRate.java
@@ -1,4 +1,4 @@
-package org.example.entity;
+package org.bolaneradarmini.entity;
 
 import com.fasterxml.jackson.annotation.JsonBackReference;
 import jakarta.persistence.*;
